edgecache, cache_volume: rename to cachevolume
